Keep screen scale when using go-to-pixel feature
Issue: #19

+++ src/ui/go_to_widget.h
@@ -37,7 +37,7 @@ class GoToWidget : public QWidget
   public:
     explicit GoToWidget(QWidget* parent = nullptr);
     void toggle_visible();
-    void set_defaults(int default_x, int default_y);
+    void set_defaults(float default_x, float default_y);
 
   Q_SIGNALS:
     void go_to_requested(int x, int y);
